refactor: rename JWTClient to AuthModel
